Refactor db instance management to factory pattern
