+++ pkg/database/migrate.go
@@ -1,238 +1,15 @@
 package database
 
 import (
-	"database/sql"
-	"errors"
 	"fmt"
-	"io"
-	glog "log"
-	"net/url"
-	"os"
-	"regexp"
-	"strings"
-	"time"
 
-	"github.com/rotisserie/eris"
-	"golang.org/x/exp/slices"
-
-	lru "github.com/hashicorp/golang-lru/v2"
-	"github.com/mattn/go-sqlite3"
 	log "github.com/sirupsen/logrus"
-	"gorm.io/driver/postgres"
 	"gorm.io/driver/sqlite"
 	"gorm.io/gorm"
 	"gorm.io/gorm/logger"
-	"gorm.io/plugin/dbresolver"
 )
 
-const (
-	SQLiteCustomDriverName = "sqlite3_custom_driver"
-)
-
-type DbInstance struct {
-	*gorm.DB
-	dsn     string
-	closers []io.Closer
-}
-
-func (db *DbInstance) Close() error {
-	for _, c := range db.closers {
-		err := c.Close()
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func (db *DbInstance) DSN() string {
-	return db.dsn
-}
-
-// DB is a global db instance.
-var DB *DbInstance = &DbInstance{}
-
-// ConnectDB will establish and return a DbInstance while also caching it in the global
-// var database.DB.
-func ConnectDB(dsn string, slowThreshold time.Duration, poolMax int, reset bool, migrate bool, artifactRoot string,
-) (*DbInstance, error) {
-	db, err := MakeDBInstance(dsn, slowThreshold, poolMax, reset, migrate, artifactRoot)
-	if err != nil {
-		return nil, err
-	}
-	// set the global DB
-	DB = db
-	return DB, nil
-}
-
-// MakeDbInstance will create a DbInstance from the parameters.
-func MakeDBInstance(
-	dsn string, slowThreshold time.Duration, poolMax int, reset bool, migrate bool, artifactRoot string,
-) (*DbInstance, error) {
-	// local db instance
-	db := DbInstance{}
-	db.dsn = dsn
-	var sourceConn gorm.Dialector
-	var replicaConn gorm.Dialector
-	u, err := url.Parse(dsn)
-	if err != nil {
-		return nil, fmt.Errorf("invalid database URL: %w", err)
-	}
-	switch u.Scheme {
-	case "postgres", "postgresql":
-		sourceConn = postgres.Open(u.String())
-	case "sqlite":
-		dbURL := *u
-		q := u.Query()
-		q.Set("_case_sensitive_like", "true")
-		q.Set("_mutex", "no")
-		if q.Get("mode") != "memory" && !(q.Has("_journal") || q.Has("_journal_mode")) {
-			q.Set("_journal", "WAL")
-		}
-		dbURL.RawQuery = q.Encode()
-
-		if reset && q.Get("mode") != "memory" {
-			file := dbURL.Host
-			if file == "" {
-				file = dbURL.Path
-			}
-			log.Infof("Removing database file %s", file)
-			if err := os.Remove(file); err != nil && !errors.Is(err, os.ErrNotExist) {
-				return nil, fmt.Errorf("failed to remove database file: %w", err)
-			}
-		}
-
-		if !slices.Contains(sql.Drivers(), SQLiteCustomDriverName) {
-			sql.Register(SQLiteCustomDriverName, &sqlite3.SQLiteDriver{
-				ConnectHook: func(conn *sqlite3.SQLiteConn) error {
-					// create LRU cache to cache regexp statements and results.
-					cache, err := lru.New[string, *regexp.Regexp](1000)
-					if err != nil {
-						return eris.Wrap(err, "error creating lru cache to cache regexp statements")
-					}
-					return conn.RegisterFunc("regexp", func(re, s string) bool {
-						result, ok := cache.Get(re)
-						if !ok {
-							result, err = regexp.Compile(re)
-							if err != nil {
-								return false
-							}
-							cache.Add(re, result)
-						}
-						return result.MatchString(s)
-					}, true)
-				},
-			})
-		}
-
-		s, err := sql.Open(SQLiteCustomDriverName, strings.Replace(dbURL.String(), "sqlite://", "file:", 1))
-		if err != nil {
-			return nil, fmt.Errorf("failed to connect to database: %w", err)
-		}
-		db.closers = append(db.closers, s)
-		s.SetMaxIdleConns(1)
-		s.SetMaxOpenConns(1)
-		s.SetConnMaxIdleTime(0)
-		s.SetConnMaxLifetime(0)
-		sourceConn = sqlite.Dialector{
-			Conn: s,
-		}
-
-		q.Set("_query_only", "true")
-		dbURL.RawQuery = q.Encode()
-		r, err := sql.Open(SQLiteCustomDriverName, strings.Replace(dbURL.String(), "sqlite://", "file:", 1))
-		if err != nil {
-			db.Close()
-			return nil, fmt.Errorf("failed to connect to database: %w", err)
-		}
-		db.closers = append(db.closers, r)
-		replicaConn = sqlite.Dialector{
-			Conn: r,
-		}
-	default:
-		return nil, fmt.Errorf("unsupported database scheme %s", u.Scheme)
-	}
-
-	logURL := *u
-	q := logURL.Query()
-	if q.Has("_key") {
-		q.Set("_key", "xxxxx")
-	}
-	logURL.RawQuery = q.Encode()
-	log.Infof("Using database %s", logURL.Redacted())
-
-	dbLogLevel := logger.Warn
-	if log.GetLevel() == log.DebugLevel {
-		dbLogLevel = logger.Info
-	}
-	db.DB, err = gorm.Open(sourceConn, &gorm.Config{
-		Logger: logger.New(
-			glog.New(
-				log.StandardLogger().WriterLevel(log.WarnLevel),
-				"",
-				0,
-			),
-			logger.Config{
-				SlowThreshold:             slowThreshold,
-				LogLevel:                  dbLogLevel,
-				IgnoreRecordNotFoundError: true,
-			},
-		),
-	})
-	if err != nil {
-		db.Close()
-		return nil, fmt.Errorf("failed to connect to database: %w", err)
-	}
-
-	if replicaConn != nil {
-		db.Use(
-			dbresolver.Register(dbresolver.Config{
-				Replicas: []gorm.Dialector{
-					replicaConn,
-				},
-			}),
-		)
-	}
-
-	if u.Scheme != "sqlite" {
-		sqlDB, _ := db.DB.DB()
-		sqlDB.SetConnMaxIdleTime(time.Minute)
-		sqlDB.SetMaxIdleConns(poolMax)
-		sqlDB.SetMaxOpenConns(poolMax)
-
-		if reset {
-			if err := resetDB(&db); err != nil {
-				db.Close()
-				return nil, err
-			}
-		}
-	}
-
-	if err := checkAndMigrateDB(&db, migrate); err != nil {
-		db.Close()
-		return nil, err
-	}
-
-	if err := createDefaultExperiment(&db, artifactRoot); err != nil {
-		db.Close()
-		return nil, err
-	}
-	return &db, nil
-}
-
-func resetDB(db *DbInstance) error {
-	switch db.Dialector.Name() {
-	case "postgres":
-		log.Info("Resetting database schema")
-		db.Exec("drop schema public cascade")
-		db.Exec("create schema public")
-	default:
-		return fmt.Errorf("unable to reset database with backend \"%s\"", db.Dialector.Name())
-	}
-	return nil
-}
-
-func checkAndMigrateDB(db *DbInstance, migrate bool) error {
+func (db *DbInstance) checkAndMigrate(migrate bool) error {
 	var alembicVersion AlembicVersion
 	var schemaVersion SchemaVersion
 	{
@@ -527,39 +304,5 @@ func checkAndMigrateDB(db *DbInstance, migrate bool) error {
 		}
 	}
 
-	return nil
-}
-
-func createDefaultExperiment(db *DbInstance, artifactRoot string) error {
-	if tx := db.First(&Experiment{}, 0); tx.Error != nil {
-		if errors.Is(tx.Error, gorm.ErrRecordNotFound) {
-			log.Info("Creating default experiment")
-			var id int32 = 0
-			ts := time.Now().UTC().UnixMilli()
-			exp := Experiment{
-				ID:             &id,
-				Name:           "Default",
-				LifecycleStage: LifecycleStageActive,
-				CreationTime: sql.NullInt64{
-					Int64: ts,
-					Valid: true,
-				},
-				LastUpdateTime: sql.NullInt64{
-					Int64: ts,
-					Valid: true,
-				},
-			}
-			if tx := db.Create(&exp); tx.Error != nil {
-				return fmt.Errorf("error creating default experiment: %s", tx.Error)
-			}
-
-			exp.ArtifactLocation = fmt.Sprintf("%s/%d", strings.TrimRight(artifactRoot, "/"), *exp.ID)
-			if tx := db.Model(&exp).Update("ArtifactLocation", exp.ArtifactLocation); tx.Error != nil {
-				return fmt.Errorf("error updating artifact_location for experiment '%s': %s", exp.Name, tx.Error)
-			}
-		} else {
-			return fmt.Errorf("unable to find default experiment: %s", tx.Error)
-		}
-	}
 	return nil
 }
